Update sample post layout
