+++ src/Homepage/styles.css
@@ -6,13 +6,15 @@ body   {
 body * {  font-family: Arial, Helvetica, sans-serif;  }
 
 @media (min-width: 768px) {
-  #sm-gamehorizon .sm-main-content-offset {
-    padding-left: 140px !important;
+  #sm-homepage {
+    /* padding-left: 140px !important; */
     background-color: gray;
   }
 }
 
 
-#sm-homepage {
-  background-color: gray;
+#sm-sample-post{
+  max-width: 75%;
+  max-height: 75%;
+  margin: auto;
 }
